chore(selectors/acl): extract 'aggregateBySubject' to a separate file [YTFRONT-5158]
